modified functions

+++ day1/CompoundInterest.java
@@ -4,10 +4,11 @@ public class CompoundInterest {
       System.out.println("Please give valid input");
       return;
     } 
-      double principal = Integer.parseInt(args[0]);
+    
+      int principal = Integer.parseInt(args[0]);
       double rateOfInterest = Integer.parseInt(args[1]);
-
       int period = Integer.parseInt(args[2]);
+
       double compoundInterest = principal * Math.pow((1 + rateOfInterest/100 ),period);
       
       System.out.println(compoundInterest - principal);
